Time: 26 ms (66.38%), Space: 49.7 MB (83.24%) - LeetHub

+++ 706-design-hashmap/706-design-hashmap.java
@@ -5,7 +5,7 @@ class MyHashMap {
     }
     
     public void put(int key, int value) {
-        val[key]=(Integer)value;
+        val[key]=value;
     }
     
     public int get(int key) {
